fix: move MacWindows.h to proper location
